LayoutTests/platform/glib/http/tests/websocket/tests/hybi/long-control-frame-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/long-control-frame' failed: Received invalid WebSocket response from the server
Test whether WebSocket rejects control frames longer than 125 bytes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
